feat: 검색결과 미리보기 기능 구현
검색결과 미리 보기 API 로직 구현
- 동작만하도록 구현한 상태
