+++ backend/src/v1/routes/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import * as auth from './auth.routes';
 import * as books from './books.routes';
-import * as bookSearchKeyword from './bookSearchKeyword.routes';
+import * as searchKeywords from './searchKeywords.routes';
 import * as lendings from './lendings.routes';
 import * as reservations from './reservations.routes';
 import * as users from './users.routes';
@@ -11,7 +11,6 @@ import * as bookInfoReviews from './bookInfoReviews.routes';
 import * as stock from './stock.routes';
 import * as tags from './tags.routes';
 import * as cursus from './cursus.routes';
-import * as searchKeywords from './search-keywords.routes';
 
 const router = Router();
 
@@ -28,7 +27,7 @@ router.use(histories.path, histories.router);
 router.use(reviews.path, reviews.router);
 router.use(histories.path, histories.router);
 router.use(bookInfoReviews.path, bookInfoReviews.router);
-router.use(bookSearchKeyword.path, bookSearchKeyword.router);
+router.use(searchKeywords.path, searchKeywords.router);
 router.use(stock.path, stock.router);
 router.use(tags.path, tags.router);
 router.use(cursus.path, cursus.router);
